+++ packages/epub_view/lib/epub_view.dart
@@ -1,1 +1,3 @@
 export 'src/epub_view.dart';
+export 'src/ui/actual_chapter.dart';
+export 'src/ui/table_of_contents.dart';