+++ packages/epub_view/lib/src/ui/table_of_contents.dart
@@ -1,0 +1,50 @@
+import 'package:epub_view/src/epub_view.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+
+class EpubReaderTableOfContents extends StatelessWidget {
+  const EpubReaderTableOfContents({
+    @required this.controller,
+    this.itemBuilder,
+    this.loader,
+    Key key,
+  }) : super(key: key);
+
+  final EpubReaderController controller;
+  final IndexedWidgetBuilder itemBuilder;
+  final Widget loader;
+
+  @override
+  Widget build(BuildContext context) => StreamBuilder<List<EpubReaderChapter>>(
+        stream: controller.tableOfContentsStream,
+        builder: (_, snapshot) {
+          Widget content;
+
+          if (snapshot.hasData) {
+            final toc = snapshot.data;
+            content = ListView.builder(
+              key: Key('$runtimeType.content'),
+              itemBuilder: itemBuilder ??
+                  (_, index) => ListTile(
+                        title: Text(toc[index].title),
+                        onTap: () =>
+                            controller.scrollTo(index: toc[index].startIndex),
+                      ),
+              itemCount: toc.length,
+            );
+          } else {
+            content = KeyedSubtree(
+              key: Key('$runtimeType.loader'),
+              child: loader ?? Center(child: CircularProgressIndicator()),
+            );
+          }
+
+          return AnimatedSwitcher(
+            duration: Duration(milliseconds: 250),
+            transitionBuilder: (Widget child, Animation<double> animation) =>
+                FadeTransition(child: child, opacity: animation),
+            child: content,
+          );
+        },
+      );
+}