+++ packages/native_pdf_view/example/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:native_pdf_view/native_pdf_view.dart';
-import 'package:native_pdf_view_example/has_support.dart';
 
 void main() => runApp(MyApp());
 
@@ -10,65 +9,55 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  int _actualPageNumber = 1;
-  PDFDocument _document;
+  int _actualPageNumber = 1, _allPagesCount = 0;
+  PdfController _pdfController;
+
+  @override
+  void initState() {
+    _pdfController = PdfController(
+      document: PdfDocument.openAsset('assets/sample.pdf'),
+    );
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    _pdfController.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) => MaterialApp(
         theme: ThemeData(primaryColor: Colors.white),
         home: Scaffold(
-          appBar: AppBar(title: Text('PDFView example')),
-          body: FutureBuilder<PDFDocument>(
-            future: _getDocument(),
-            builder: (_, snapshot) {
-              if (snapshot.hasData) {
-                return Stack(
-                  children: <Widget>[
-                    PDFView(
-                      document: snapshot.data,
-                      onPageChanged: (page) {
-                        setState(() {
-                          _actualPageNumber = page;
-                        });
-                      },
-                    ),
-                    Align(
-                      alignment: Alignment.center,
-                      child: Text(
-                        '$_actualPageNumber/${snapshot.data.pagesCount}',
-                        style: TextStyle(fontSize: 34),
-                      ),
-                    )
-                  ],
-                );
-              }
-
-              if (snapshot.hasError) {
-                return Center(
-                  child: Text(
-                    'PDF Rendering does not '
-                    'support on the system of this version',
-                  ),
-                );
-              }
-
-              return Center(child: CircularProgressIndicator());
+          appBar: AppBar(
+            title: Text('PdfView example'),
+            actions: <Widget>[
+              Container(
+                margin: EdgeInsets.only(right: 16),
+                alignment: Alignment.center,
+                child: Text(
+                  '$_actualPageNumber/$_allPagesCount',
+                  style: TextStyle(fontSize: 22),
+                ),
+              ),
+            ],
+          ),
+          body: PdfView(
+            documentLoader: Center(child: CircularProgressIndicator()),
+            pageLoader: Center(child: CircularProgressIndicator()),
+            controller: _pdfController,
+            onDocumentLoaded: (document) {
+              setState(() {
+                _allPagesCount = document.pagesCount;
+              });
+            },
+            onPageChanged: (page) {
+              setState(() {
+                _actualPageNumber = page;
+              });
             },
           ),
         ),
       );
-
-  Future<PDFDocument> _getDocument() async {
-    if (_document != null) {
-      return _document;
-    }
-    if (await hasSupport()) {
-      return _document = await PDFDocument.openAsset('assets/sample.pdf');
-    } else {
-      throw Exception(
-        'PDF Rendering does not '
-        'support on the system of this version',
-      );
-    }
-  }
 }